separate system-probe marshal and unmarshal code
