+++ cmd/system-probe/modules/network_tracer_test.go
@@ -16,7 +16,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/DataDog/datadog-agent/pkg/network"
-	"github.com/DataDog/datadog-agent/pkg/network/encoding"
+	"github.com/DataDog/datadog-agent/pkg/network/encoding/marshal"
+	"github.com/DataDog/datadog-agent/pkg/process/encoding"
 	"github.com/DataDog/datadog-agent/pkg/process/util"
 )
 
@@ -59,10 +60,10 @@ func TestDecode(t *testing.T) {
 		},
 	}
 
-	marshaller := encoding.GetMarshaler(encoding.ContentTypeJSON)
+	marshaller := marshal.GetMarshaler(encoding.ContentTypeJSON)
 	ostream := bytes.NewBuffer(nil)
 
-	connectionsModeler := encoding.NewConnectionsModeler(in)
+	connectionsModeler := marshal.NewConnectionsModeler(in)
 	defer connectionsModeler.Close()
 
 	err := marshaller.Marshal(in, ostream, connectionsModeler)
