+++ pkg/network/encoding/marshal/http2.go
@@ -3,38 +3,39 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package encoding
+package marshal
 
 import (
 	"bytes"
 	"io"
 
 	model "github.com/DataDog/agent-payload/v5/process"
+	"github.com/gogo/protobuf/proto"
+
 	"github.com/DataDog/datadog-agent/pkg/network"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/http"
 	"github.com/DataDog/datadog-agent/pkg/network/types"
-	"github.com/gogo/protobuf/proto"
 )
 
-type httpEncoder struct {
-	httpAggregationsBuilder *model.HTTPAggregationsBuilder
-	byConnection            *USMConnectionIndex[http.Key, *http.RequestStats]
+type http2Encoder struct {
+	http2AggregationsBuilder *model.HTTP2AggregationsBuilder
+	byConnection             *USMConnectionIndex[http.Key, *http.RequestStats]
 }
 
-func newHTTPEncoder(httpPayloads map[http.Key]*http.RequestStats) *httpEncoder {
-	if len(httpPayloads) == 0 {
+func newHTTP2Encoder(http2Payloads map[http.Key]*http.RequestStats) *http2Encoder {
+	if len(http2Payloads) == 0 {
 		return nil
 	}
 
-	return &httpEncoder{
-		httpAggregationsBuilder: model.NewHTTPAggregationsBuilder(nil),
-		byConnection: GroupByConnection("http", httpPayloads, func(key http.Key) types.ConnectionKey {
+	return &http2Encoder{
+		byConnection: GroupByConnection("http2", http2Payloads, func(key http.Key) types.ConnectionKey {
 			return key.ConnectionKey
 		}),
+		http2AggregationsBuilder: model.NewHTTP2AggregationsBuilder(nil),
 	}
 }
 
-func (e *httpEncoder) GetHTTPAggregationsAndTags(c network.ConnectionStats, builder *model.ConnectionBuilder) (uint64, map[string]struct{}) {
+func (e *http2Encoder) WriteHTTP2AggregationsAndTags(c network.ConnectionStats, builder *model.ConnectionBuilder) (uint64, map[string]struct{}) {
 	if e == nil {
 		return 0, nil
 	}
@@ -49,34 +50,33 @@ func (e *httpEncoder) GetHTTPAggregationsAndTags(c network.ConnectionStats, buil
 		dynamicTags map[string]struct{}
 	)
 
-	builder.SetHttpAggregations(func(b *bytes.Buffer) {
+	builder.SetHttp2Aggregations(func(b *bytes.Buffer) {
 		staticTags, dynamicTags = e.encodeData(connectionData, b)
 	})
 	return staticTags, dynamicTags
 }
 
-func (e *httpEncoder) encodeData(connectionData *USMConnectionData[http.Key, *http.RequestStats], w io.Writer) (uint64, map[string]struct{}) {
+func (e *http2Encoder) encodeData(connectionData *USMConnectionData[http.Key, *http.RequestStats], w io.Writer) (uint64, map[string]struct{}) {
 	var staticTags uint64
 	dynamicTags := make(map[string]struct{})
-	e.httpAggregationsBuilder.Reset(w)
+	e.http2AggregationsBuilder.Reset(w)
 
 	for _, kvPair := range connectionData.Data {
-		e.httpAggregationsBuilder.AddEndpointAggregations(func(httpStatsBuilder *model.HTTPStatsBuilder) {
+		e.http2AggregationsBuilder.AddEndpointAggregations(func(http2StatsBuilder *model.HTTPStatsBuilder) {
 			key := kvPair.Key
 			stats := kvPair.Value
 
-			httpStatsBuilder.SetPath(key.Path.Content.Get())
-			httpStatsBuilder.SetFullPath(key.Path.FullPath)
-			httpStatsBuilder.SetMethod(uint64(model.HTTPMethod(key.Method)))
+			http2StatsBuilder.SetPath(key.Path.Content.Get())
+			http2StatsBuilder.SetFullPath(key.Path.FullPath)
+			http2StatsBuilder.SetMethod(uint64(model.HTTPMethod(key.Method)))
 
 			for code, stats := range stats.Data {
-				httpStatsBuilder.AddStatsByStatusCode(func(w *model.HTTPStats_StatsByStatusCodeEntryBuilder) {
+				http2StatsBuilder.AddStatsByStatusCode(func(w *model.HTTPStats_StatsByStatusCodeEntryBuilder) {
 					w.SetKey(int32(code))
 					w.SetValue(func(w *model.HTTPStats_DataBuilder) {
 						w.SetCount(uint32(stats.Count))
 						if latencies := stats.Latencies; latencies != nil {
 
-							// TODO: can we get a streaming marshaller for latencies?
 							blob, _ := proto.Marshal(latencies.ToProto())
 							w.SetLatencies(func(b *bytes.Buffer) {
 								b.Write(blob)
@@ -93,12 +93,12 @@ func (e *httpEncoder) encodeData(connectionData *USMConnectionData[http.Key, *ht
 				}
 			}
 		})
-
 	}
+
 	return staticTags, dynamicTags
 }
 
-func (e *httpEncoder) Close() {
+func (e *http2Encoder) Close() {
 	if e == nil {
 		return
 	}
