+++ pkg/network/encoding/marshal/marshaler.go
@@ -1,0 +1,39 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package marshal
+
+import (
+	"io"
+	"strings"
+
+	"github.com/gogo/protobuf/jsonpb"
+
+	"github.com/DataDog/datadog-agent/pkg/network"
+)
+
+var (
+	pSerializer = protoSerializer{}
+	jSerializer = jsonSerializer{
+		marshaller: jsonpb.Marshaler{
+			EmitDefaults: true,
+		},
+	}
+)
+
+// Marshaler is an interface implemented by all Connections serializers
+type Marshaler interface {
+	Marshal(conns *network.Connections, writer io.Writer, connsModeler *ConnectionsModeler) error
+	ContentType() string
+}
+
+// GetMarshaler returns the appropriate Marshaler based on the given accept header
+func GetMarshaler(accept string) Marshaler {
+	if strings.Contains(accept, ContentTypeProtobuf) {
+		return pSerializer
+	}
+
+	return jSerializer
+}
