+++ pkg/network/encoding/marshal/modeler.go
@@ -3,14 +3,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package encoding
+package marshal
 
 import (
-	"io"
-	"strings"
 	"sync"
-
-	"github.com/gogo/protobuf/jsonpb"
 
 	model "github.com/DataDog/agent-payload/v5/process"
 
@@ -19,36 +15,9 @@ import (
 )
 
 var (
-	pSerializer = protoSerializer{}
-	jSerializer = jsonSerializer{
-		marshaller: jsonpb.Marshaler{
-			EmitDefaults: true,
-		},
-	}
-
 	cfgOnce  = sync.Once{}
 	agentCfg *model.AgentConfiguration
 )
-
-// Marshaler is an interface implemented by all Connections serializers
-type Marshaler interface {
-	Marshal(conns *network.Connections, writer io.Writer, connsModeler *ConnectionsModeler) error
-	ContentType() string
-}
-
-// Unmarshaler is an interface implemented by all Connections deserializers
-type Unmarshaler interface {
-	Unmarshal([]byte) (*model.Connections, error)
-}
-
-// GetMarshaler returns the appropriate Marshaler based on the given accept header
-func GetMarshaler(accept string) Marshaler {
-	if strings.Contains(accept, ContentTypeProtobuf) {
-		return pSerializer
-	}
-
-	return jSerializer
-}
 
 // ConnectionsModeler contains all the necessary structs for modeling a connection.
 type ConnectionsModeler struct {
@@ -84,15 +53,6 @@ func (c *ConnectionsModeler) Close() {
 	c.httpEncoder.Close()
 	c.http2Encoder.Close()
 	c.kafkaEncoder.Close()
-}
-
-// GetUnmarshaler returns the appropriate Unmarshaler based on the given content type
-func GetUnmarshaler(ctype string) Unmarshaler {
-	if strings.Contains(ctype, ContentTypeProtobuf) {
-		return pSerializer
-	}
-
-	return jSerializer
 }
 
 func (c *ConnectionsModeler) modelConnections(builder *model.ConnectionsBuilder, conns *network.Connections) {
