+++ pkg/network/encoding/marshal/usm_lookup_windows.go
@@ -3,9 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build !windows
+//go:build windows
 
-package encoding
+package marshal
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/network"
@@ -13,25 +13,12 @@ import (
 )
 
 // USMLookup determines the strategy for associating a given connection to USM
-// In the context of Linux we may perform up to 4 lookups as described below
 func USMLookup[K comparable, V any](c network.ConnectionStats, data map[types.ConnectionKey]*USMConnectionData[K, V]) *USMConnectionData[K, V] {
-	var connectionData *USMConnectionData[K, V]
-
-	// WithKey will attempt 4 lookups in total
-	// 1) (A, B)
-	// 2) (B, A)
-	// 3) (translated(A), translated(B))
-	// 3) (translated(B), translated(A))
-	// The callback API is used to avoid allocating a slice of all pre-computed keys
-	network.WithKey(c, func(key types.ConnectionKey) (stopIteration bool) {
-		val, ok := data[key]
-		if !ok {
-			return false
+	for _, key := range network.ConnectionKeysFromConnectionStats(c) {
+		if v, ok := data[key]; ok {
+			return v
 		}
+	}
 
-		connectionData = val
-		return true
-	})
-
-	return connectionData
+	return nil
 }
