+++ pkg/network/encoding/marshal/usm_payload_telemetry.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package encoding
+package marshal
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/network"
