+++ pkg/network/encoding/unmarshal/protobuf.go
@@ -1,0 +1,28 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package unmarshal
+
+import (
+	model "github.com/DataDog/agent-payload/v5/process"
+	"github.com/gogo/protobuf/proto"
+)
+
+// ContentTypeProtobuf holds the HTML content-type of a Protobuf payload
+const ContentTypeProtobuf = "application/protobuf"
+
+type protoSerializer struct{}
+
+func (protoSerializer) Unmarshal(blob []byte) (*model.Connections, error) {
+	conns := new(model.Connections)
+	if err := proto.Unmarshal(blob, conns); err != nil {
+		return nil, err
+	}
+	return conns, nil
+}
+
+func (p protoSerializer) ContentType() string {
+	return ContentTypeProtobuf
+}
